+++ src/Test/index.ts
@@ -6,9 +6,7 @@ import "@babylonjs/core/Engines/WebGPU/Extensions/engine.renderTarget";
 import "@babylonjs/core/Engines/WebGPU/Extensions/engine.renderTargetTexture";
 
 import { BaseRuntime } from "./baseRuntime";
-import { SceneBuilder } from "./Scene/constraintTestScene";
-
-await new Promise(resolve => window.onload = resolve);
+import { SceneBuilder } from "./Scene/wasmPhysicsTestScene";
 
 const canvas = document.createElement("canvas");
 canvas.style.width = "100%";